requirements.txt
Django>=4.2.16
channels==4.0.0
channels-redis==4.1.0
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
django-environ==0.11.2
whitenoise==6.6.0
Pillow==10.2.0
daphne==4.0.0
docker==7.0.0
kombu==5.3.4
vine==5.1.0
asgiref>=3.8.1
django-crispy-forms>=2.3
crispy-tailwind==1.0.3
